主播管理

+++ qfstream/src/main/java/com/rongtuoyouxuan/chatlive/stream/view/msgspan/BaseMessageSpanMatcher.kt
@@ -82,7 +82,7 @@ abstract class BaseMessageSpanMatcher(context: Context) : IMessageSpanMatcher {
      */
     open fun addNickName(textView: TextView, common:BaseRoomMessage, spanString: SpannableStringBuilder, nameColor: Int, addMaohao: Boolean) {
         addTextSpan(spanString, textView, common.userName + if (addMaohao) ": " else " ", mContext.resources.getColor(R.color.c_stream_msg_nick_name), ClickiTextSpan.SpanClick {
-            LiveRoomHelper.openUserCardVM.post(common.userId.toLong())
+            LiveRoomHelper.openUserCardVM.post(common.userId)
             ULog.d("clll", "addNickName")
         })
     }
@@ -93,7 +93,7 @@ abstract class BaseMessageSpanMatcher(context: Context) : IMessageSpanMatcher {
     open fun addNickName(textView: TextView, common:BaseRoomMessage, spanString: SpannableStringBuilder, addMaohao: Boolean) {
         addTextSpan(spanString, textView, common.userName + if (addMaohao) ": " else " ", mContext.resources.getColor(
             R.color.c_stream_msg_common), ClickiTextSpan.SpanClick {
-            LiveRoomHelper.openUserCardVM.post(common.userId.toLong())
+            LiveRoomHelper.openUserCardVM.post(common.userId)
         })
     }
